requirements.txt
python-dotenv
openai==1.3.4
tiktoken
click
PyPDF2 
requests==2.31.0
sqlite3==2.6.0

# rest api
flask==2.3.2
flask-cors==3.0.10
singleton-decorator==1.0.0

# internet support
bs4
selenium

# vectorDB
chromadb==0.4.17
langchain==0.0.339
